% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_util.R
\name{all_1d_subencodings}
\alias{all_1d_subencodings}
\title{Get all 1d sub-encodings}
\usage{
all_1d_subencodings(e, transformation = NULL, ...)
}
\arguments{
\item{e}{An \code{\link{encoding}}}

\item{transformation}{A transformation to apply to each subencoding}

\item{...}{Additional arguments to \code{transformation}}
}
\value{
A \code{\link{dplyr::tbl}} containing one encoding for each
dimension of \code{e}, with a column \code{fname} labelling the
subencodings by the \code{fname} in the original encoding
}
\description{
Get all 1d sub-encodings
}

