% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{as.tbl.encoding}
\alias{as.tbl.encoding}
\title{Convert encoding to a \code{\link{dplyr::tbl}}}
\usage{
as.tbl.encoding(e)
}
\arguments{
\item{e}{An \code{\link{encoding}}}
}
\value{
A \code{\link{dplyr::tbl}} containing all the rows of the encoding
matrix with an additional column called \code{label} that
contains the corresponding labels for those rows
}
\description{
Convert encoding to a \code{\link{dplyr::tbl}}
}

