% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encoding.matrix}
\alias{encoding.matrix}
\title{Create \code{encoding} object}
\usage{
\method{encoding}{matrix}(m, label, fnames, transformation = NULL, ...)
}
\arguments{
\item{m}{a matrix}

\item{label}{a string vector}

\item{transformation}{a transformation that will be applied to \code{m}}

\item{...}{Additional arguments to \code{transformation}}

\item{fname}{a string vector}
}
\value{
See \code{\link{encoding}}
}
\description{
Create \code{encoding} object from a matrix
with \code{label} corresponding \code{label}, \code{m}
corresponding \code{m}, and \code{fname} corresponding to \code{fname}
}

