% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_test.R
\name{generic_test}
\alias{generic_test}
\title{Conduct a class test}
\usage{
generic_test(encoding, test_classes_f, fit_and_predict_fn)
}
\arguments{
\item{encoding}{an \code{\link{encoding}} object}

\item{fit_and_predict_fn}{a function of three arguments: \code{x_tr}
(an N by d array of training vectors), \code{y_tr} (a vector of training labels),
\code{x_te} (a matrix or vector containing a point to classify), returning a
predicted label.}

\item{test_classes}{a data frame specifying the two classes: it contains a column
column \code{label} giving the label of the element,
and a column \code{value} giving the classification label}
}
\value{
a \code{\link{dplyr::tbl}} containing classification scores,
having at least one column,  \code{avg_loo}, giving the average
leave-one-out classification score
}
\description{
Conducts a class separation test on a single feature as defined in
\code{test_classes_f}
}

