% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{get.encoding}
\alias{get.encoding}
\title{Get encoding corresponding to a label}
\usage{
get.encoding(e, label)
}
\arguments{
\item{e}{An \code{\link{encoding}}}

\item{label}{A character string containing a label}
}
\value{
All the rows of the encoding matrix that have \code{label} as their
label (extra dimensions will be dropped to make it a vector if there is only one row)
}
\description{
Get encoding corresponding to a label
}

