% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_constant.R
\name{non_constant}
\alias{non_constant}
\title{Finds non-constant dimensions}
\usage{
non_constant(m)
}
\arguments{
\item{m}{A matrix(-like)}
}
\value{
indices of non-constant dimensions of \code{m}
}
\description{
Return indexes of non-constant dimensions
}

