% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mptest.R
\name{pair_labels}
\alias{pair_labels}
\title{Label pairs}
\usage{
pair_labels(lab1, lab2, collapse_main = "::", collapse_sub = "_")
}
\arguments{
\item{lab1}{Labels of left elements}

\item{lab2}{Labels of right elements}

\item{collapse_main}{Separating string}

\item{collapse_sub}{String to replace \code{collapse_main} with if it appears
in lab1 or lab2 already}
}
\value{
A character vector containing labels for the pairs
}
\description{
Label pairs
}

