% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pca}
\alias{pca}
\title{Apply PCA transform}
\usage{
pca(m, k = NULL)
}
\arguments{
\item{m}{A matrix}

\item{k}{Fixed number of dimensions (principal components) to return; defaults to
no dimensionality reduction if \code{k} is \code{NULL} - reducing by variance explained is not supported}
}
\value{
m decorrelated by PCA-rotation
}
\description{
Applies PCA and reduces to \code{k} dimensions; works by calling prcomp with z-scoring first
}

