% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_util.R
\name{pca_variance_proportion}
\alias{pca_variance_proportion}
\title{Proportion of variance explained by PCA to \code{k} #' dimensions}
\usage{
pca_variance_proportion(m, k)
}
\arguments{
\item{m}{A matrix}

\item{k}{Number of dimensions}
}
\value{
Sum of the singular values up to \code{k}
}
\description{
Proportion of variance explained by PCA to \code{k} #' dimensions
}

