% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_util.R
\name{pcasds_variance_proportion}
\alias{pcasds_variance_proportion}
\title{Proportion of variance explained by set of PCA sd's}
\usage{
pcasds_variance_proportion(sds)
}
\arguments{
\item{sds}{A vector of sds as output by \code{\link{prcomp}}}
}
\value{
Sum of the singular values up to \code{k}
}
\description{
Proportion of variance explained by set of PCA sd's
}

