% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_test.R
\name{svm_class_test}
\alias{svm_class_test}
\title{Conduct a class test using SVM}
\usage{
svm_class_test(encoding, test_classes_f, parallel = T)
}
\arguments{
\item{encoding}{an \code{\link{encoding}} object}

\item{test_classes_f}{a data frame specifying the two classes: it contains a
column \code{label} giving the label of the element,
and a column \code{value} giving the classification label}
}
\value{
a \code{\link{dplyr::tbl}} containing classification scores,
having at least one column,  \code{avg_loo}, giving the average
leave-one-out classification score
}
\description{
Conducts a leave-one-out class separation test on a single
feature as defined in \code{test_classes_f} using SVM
}

