% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{whiten_zca}
\alias{whiten_zca}
\title{Apply ZCA transform}
\usage{
whiten_zca(m, k = NULL)
}
\arguments{
\item{m}{A matrix}

\item{k}{Fixed number of dimensions (principal components) to return}

\item{prop_var}{Select dimension by proportion of explained variance (need to explain
at least \code{prop_var}); overrides \code{k} if set}
}
\value{
m whitened by PCA-rotation and z-scoring, and then re-rotated back to its original dimensions
}
\description{
Performs PCA on \code{m}, divides out the variance, and then
undoes the original PCA rotation
}

