% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{zscore}
\alias{zscore}
\title{Apply z-score transform}
\usage{
zscore(m, k = NULL)
}
\arguments{
\item{m}{A matrix}

\item{k}{Fixed number of dimensions to return; defaults to
no dimensionality reduction if \code{k} is \code{NULL}}
}
\value{
z-scored \code{m}
}
\description{
Centers and divides out the variance from \code{m}
}

