% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{get.encoding}
\alias{get.encoding}
\title{Get encoding corresponding to a set of labels}
\usage{
get.encoding(e, labels)
}
\arguments{
\item{e}{An \code{\link{encoding}}}

\item{labels}{A character string containing a set of labels}
}
\value{
A \code{\link{dplyr::tbl}} containing all the rows of the encoding
matrix that have a
label in \code{labels}, with an additional column called \code{label} that
contains the corresponding labels for those rows
}
\description{
Get encoding corresponding to a set of labels
}

