% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pca}
\alias{pca}
\title{Apply PCA transform}
\usage{
pca(m, k = NULL, prop_var = NULL)
}
\arguments{
\item{m}{A matrix}

\item{k}{Fixed number of dimensions (principal components) to return}

\item{prop_var}{Select dimension by proportion of explained variance (need to explain
at least \code{prop_var}); overrides \code{k} if set}
}
\value{
m decorrelated by PCA-rotation
}
\description{
Applies PCA and reduces to \code{k} dimensions; works by calling prcomp with z-scoring first
}

