% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{whiten_pca}
\alias{whiten_pca}
\title{Apply whitening PCA transform}
\usage{
whiten_pca(m, k = NULL)
}
\arguments{
\item{m}{A matrix}

\item{k}{Fixed number of dimensions (principal components) to return; defaults to
no dimensionality reduction if \code{k} is \code{NULL}}
}
\value{
m whitened by PCA-rotation and z-scoring
}
\description{
Applies PCA and reduces to \code{k} dimensions, then divides out the variance
}

